+++ agentsr/__init__.py
@@ -6,5 +6,20 @@ from .controller import (
     derive_agent_plan,
     observation_consistency_project,
 )
+from .reranker import CandidateScore, ScoreWeights, score_candidates
+from .token_editor import MeissonicTokenEditor, TokenRefineResult
+from .token_masks import TokenMaskSet, build_initial_token_masks
 
-__all__ = ["AgentPlan", "build_refinement_assets", "derive_agent_plan", "observation_consistency_project"]
+__all__ = [
+    "AgentPlan",
+    "CandidateScore",
+    "MeissonicTokenEditor",
+    "ScoreWeights",
+    "TokenMaskSet",
+    "TokenRefineResult",
+    "build_initial_token_masks",
+    "build_refinement_assets",
+    "derive_agent_plan",
+    "observation_consistency_project",
+    "score_candidates",
+]